decompose main.py - UserCreateDialog.py - UserEditDialog.py
